+++ src/sdlang/ast.d
@@ -854,8 +854,7 @@ class Tag
 			return NamespaceAccess(
 				namespace,
 				AttributeRange(tag, namespace, isMaybe),
-				TagRange(tag, namespace, isMaybe),
-				tag
+				TagRange(tag, namespace, isMaybe)
 			);
 		}
 		
@@ -870,8 +869,7 @@ class Tag
 			return NamespaceAccess(
 				namespace,
 				AttributeRange(tag, namespace, isMaybe),
-				TagRange(tag, namespace, isMaybe),
-				tag
+				TagRange(tag, namespace, isMaybe)
 			);
 		}
 		
@@ -895,126 +893,6 @@ class Tag
 		string name;
 		AttributeRange attributes;
 		TagRange tags;
-
-		private Tag containingTag;
-		
-		/++
-		Lookup a tag by name, and reterive a value of type T from it. Useful
-		if you only evpect one value of type T from a given tag. Only looks
-		for immediate child tags of 'this', doesn't search recursively.
-		
-		If there are multiple tags by the chosen name, the *last tag* will always
-		be chosen. That is, this function considers later tags with the same name
-		to override previous ones.
-		
-		If the chosen tag has multiple values, the *first* value matching the
-		requested type will be returned. Ie, Extra values in the tag are ignored.
-		
-		If the chosen tag doesn't have a value of the requested type, and you
-		DID provide a default value, the default value is returned.
-		
-		If the chosen tag doesn't have a value of the requested type, and you
-		DIDN'T provide a default value, a SDLangRangeException will be thrown.
-		+/
-		T getValue(T)(string tagName) if(isValueType!T)
-		{
-			return containingTag.getValueImpl!T(this.name, tagName);
-		}
-		
-		///ditto
-		T getValue(T)(string tagName, T defaultValue) if(isValueType!T)
-		{
-			return containingTag.getValueImpl!T(this.name, tagName, defaultValue);
-		}
-		
-		///
-		@("Tag.NamespaceAccess.getValue")
-		unittest
-		{
-			import std.exception;
-			import sdlang.parser;
-			
-			auto root = parseSource(`
-				ns1:foo 1 "a" 2 "b"
-				ns1:foo 3 "c" 4 "d"
-				ns2:foo 11 "aa" 22 "bb"
-				ns2:foo 33 "cc" 44 "dd"
-				
-				ns1:bar "hi"
-				ns1:bar 379  // getValue considers this to override the first bar
-			`);
-			assert( root.namespaces["ns1"].getValue!int("foo") == 3 );
-			assert( root.all.getValue!int("foo") == 33 );
-
-			assert( root.namespaces["ns1"].getValue!string("foo") == "c" );
-			assert( root.all.getValue!string("foo") == "cc" );
-			
-			// The last "bar" tag doesn't have an int (only the first "bar" tag does)
-			assertThrown!SDLangRangeException( root.all.getValue!string("bar") );
-			assert( root.all.getValue!string("bar", "Default") == "Default" );
-		}
-
-		/++
-		Lookup a tag and attribute by name, and reterive a value of type T from it.
-		Useful if you only evpect one attribute of type T from a given tag and
-		attribute name. Only looks for immediate child tags of 'this', doesn't
-		search recursively.
-		
-		If there are multiple tags by the chosen name, the *last tag* will always
-		be chosen. That is, this function considers later tags with the same name
-		to override previous ones.
-		
-		If the chosen tag has multiple attributes with the same same, the *first*
-		such attribute matching the requested type will be returned. Ie, Extra
-		attributes in the tag are ignored.
-		
-		If the chosen tag doesn't have an attribute of the requested type, and you
-		DID provide a default value, the default value is returned.
-		
-		If the chosen tag doesn't have an attribute of the requested type, and you
-		DIDN'T provide a default value, a SDLangRangeException will be thrown.
-		
-		Note: Currently, this can only be used to lookup attributes in the default
-		namespace. It can, however, lookup attributes on TAGS that are in a
-		particular namespace.
-		+/
-		T getAttribute(T)(string tagName, string attrName) if(isValueType!T)
-		{
-			return containingTag.getAttributeImpl!T(this.name, tagName, "", attrName);
-		}
-		
-		///ditto
-		T getAttribute(T)(string tagName, string attrName, T defaultValue) if(isValueType!T)
-		{
-			return containingTag.getAttributeImpl!T(this.name, tagName, "", attrName, defaultValue);
-		}
-
-		///
-		@("Tag.NamespaceAccess.getAttribute")
-		unittest
-		{
-			import std.exception;
-			import sdlang.parser;
-			
-			auto root = parseSource(`
-				ns1:foo X=1 X="a" X=2 X="b"
-				ns1:foo X=3 X="c" X=4 X="d"
-				ns2:foo X=11 X="aa" X=22 X="bb"
-				ns2:foo X=33 X="cc" X=44 X="dd"
-				
-				ns1:bar X="hi"
-				ns1:bar X=379  // getAttribute considers this to override the first bar
-			`);
-			assert( root.namespaces["ns1"].getAttribute!int("foo", "X") == 3 );
-			assert( root.all.getAttribute!int("foo", "X") == 33 );
-			
-			assert( root.namespaces["ns1"].getAttribute!string("foo", "X") == "c" );
-			assert( root.all.getAttribute!string("foo", "X") == "cc" );
-			
-			// The last "bar" tag doesn't have an int attribute named "X" (only the first "bar" tag does)
-			assertThrown!SDLangRangeException( root.all.getAttribute!string("bar", "X") );
-			assert( root.all.getAttribute!string("bar", "X", "Default") == "Default" );
-		}
 	}
 
 	alias MemberRange!(Attribute, "allAttributes", "attributeIndicies", "_attributes") AttributeRange;
@@ -1048,8 +926,7 @@ class Tag
 		return NamespaceAccess(
 			"*",
 			AttributeRange(this, "*", false),
-			TagRange(this, "*", false),
-			this
+			TagRange(this, "*", false)
 		);
 	}
 
@@ -1082,21 +959,8 @@ class Tag
 			return NamespaceAccess(
 				"*",
 				AttributeRange(tag, "*", true),
-				TagRange(tag, "*", true),
-				tag
+				TagRange(tag, "*", true)
 			);
-		}
-		
-		/// See the documentation for Tag.getValue
-		T getValue(T)(string tagName) if(isValueType!T)
-		{
-			return getValueImpl!T("", tagName);
-		}
-		
-		///ditto
-		T getValue(T)(string tagName, T defaultValue) if(isValueType!T)
-		{
-			return getValueImpl!T("", tagName, defaultValue);
 		}
 	}
 	
@@ -1113,6 +977,9 @@ class Tag
 	if you only evpect one value of type T from a given tag. Only looks
 	for immediate child tags of 'this', doesn't search recursively.
 	
+	The name can optionally include a namespace, as in "namespace:name".
+	Or, you can search all namespaces using "*:name".
+	
 	If there are multiple tags by the chosen name, the *last tag* will always
 	be chosen. That is, this function considers later tags with the same name
 	to override previous ones.
@@ -1128,13 +995,15 @@ class Tag
 	+/
 	T getValue(T)(string tagName) if(isValueType!T)
 	{
-		return getValueImpl!T("", tagName);
+		auto splitName = FullName.split(tagName);
+		return getValueImpl!T(splitName.namespace, splitName.name);
 	}
 	
 	///ditto
 	T getValue(T)(string tagName, T defaultValue) if(isValueType!T)
 	{
-		return getValueImpl!T("", tagName, defaultValue);
+		auto splitName = FullName.split(tagName);
+		return getValueImpl!T(splitName.namespace, splitName.name, defaultValue);
 	}
 	
 	///
@@ -1157,6 +1026,26 @@ class Tag
 		// The last "bar" tag doesn't have an int (only the first "bar" tag does)
 		assertThrown!SDLangRangeException( root.getValue!string("bar") );
 		assert( root.getValue!string("bar", "Default") == "Default" );
+
+		// Using namespaces:
+		root = parseSource(`
+			ns1:foo 1 "a" 2 "b"
+			ns1:foo 3 "c" 4 "d"
+			ns2:foo 11 "aa" 22 "bb"
+			ns2:foo 33 "cc" 44 "dd"
+			
+			ns1:bar "hi"
+			ns1:bar 379  // getValue considers this to override the first bar
+		`);
+		assert( root.getValue!int("ns1:foo") == 3  );
+		assert( root.getValue!int("*:foo"  ) == 33 ); // Search all namespaces
+
+		assert( root.getValue!string("ns1:foo") == "c"  );
+		assert( root.getValue!string("*:foo"  ) == "cc" ); // Search all namespaces
+		
+		// The last "bar" tag doesn't have an int (only the first "bar" tag does)
+		assertThrown!SDLangRangeException( root.getValue!string("*:bar") );
+		assert( root.getValue!string("*:bar", "Default") == "Default" );
 	}
 	
 	private T getValueImpl(T)(string namespace, string tagName) if(isValueType!T)
@@ -1179,12 +1068,7 @@ class Tag
 			if(useDefaultValue)
 				return defaultValue;
 			else
-			{
-				if(namespace == "*")
-					throw new SDLangRangeException("Can't find tag '"~tagName~"' in any namespace");
-				else
-					throw new SDLangRangeException("Can't find tag '"~tagName~"' in namespace '"~namespace~"'");
-			}
+				throw new SDLangRangeException("Can't find tag matching '"~FullName.combine(namespace, tagName)~"'");
 		}
 
 		foreach(value; _tags[namespace][tagName][$-1].values)
@@ -1197,20 +1081,10 @@ class Tag
 			return defaultValue;
 		else
 		{
-			if(namespace == "*")
-			{
-				throw new SDLangRangeException(
-					"The last tag named '"~tagName~"' in all namespaces doesn't "~
-					"have a value of type "~T.stringof
-				);
-			}
-			else
-			{
-				throw new SDLangRangeException(
-					"The last tag named '"~tagName~"' in namespace '"~namespace~"' doesn't "~
-					"have a value of type "~T.stringof
-				);
-			}
+			throw new SDLangRangeException(
+				"The last tag matching '"~FullName.combine(namespace, tagName)~"' doesn't "~
+				"have a value of type "~T.stringof
+			);
 		}
 	}
 
@@ -1220,6 +1094,9 @@ class Tag
 	attribute name. Only looks for immediate child tags of 'this', doesn't
 	search recursively.
 	
+	The tag and attributes names can optionally include a namespace, as in
+	"namespace:name". Or, you can search all namespaces using "*:name".
+
 	If there are multiple tags by the chosen name, the *last tag* will always
 	be chosen. That is, this function considers later tags with the same name
 	to override previous ones.
@@ -1236,13 +1113,23 @@ class Tag
 	+/
 	T getAttribute(T)(string tagName, string attrName) if(isValueType!T)
 	{
-		return getAttributeImpl!T("", tagName, "", attrName);
+		auto splitTagName = FullName.split(tagName);
+		auto splitAttrName = FullName.split(attrName);
+		return getAttributeImpl!T(
+			splitTagName.namespace, splitTagName.name,
+			splitAttrName.namespace, splitAttrName.name
+		);
 	}
 	
 	///ditto
 	T getAttribute(T)(string tagName, string attrName, T defaultValue) if(isValueType!T)
 	{
-		return getAttributeImpl!T("", tagName, "", attrName, defaultValue);
+		auto splitTagName = FullName.split(tagName);
+		auto splitAttrName = FullName.split(attrName);
+		return getAttributeImpl!T(
+			splitTagName.namespace, splitTagName.name,
+			splitAttrName.namespace, splitAttrName.name, defaultValue
+		);
 	}
 	
 	///
@@ -1265,40 +1152,57 @@ class Tag
 		// The last "bar" tag doesn't have an int attribute named "X" (only the first "bar" tag does)
 		assertThrown!SDLangRangeException( root.getAttribute!string("bar", "X") );
 		assert( root.getAttribute!string("bar", "X", "Default") == "Default" );
+		
+
+		// Using namespaces:
+		root = parseSource(`
+			ns1:foo X=1 X="a" X=2 X="b"
+			ns1:foo X=3 X="c" X=4 X="d"
+			ns2:foo X=11 X="aa" X=22 X="bb"
+			ns2:foo X=33 X="cc" X=44 X="dd"
+			
+			ns1:bar attrNS:X="hi"
+			ns1:bar attrNS:X=379  // getAttribute considers this to override the first bar
+		`);
+		assert( root.getAttribute!int("ns1:foo", "X") == 3  );
+		assert( root.getAttribute!int("*:foo",   "X") == 33 ); // Search all namespaces
+		
+		assert( root.getAttribute!string("ns1:foo", "X") == "c"  );
+		assert( root.getAttribute!string("*:foo",   "X") == "cc" ); // Search all namespaces
+		
+		// The last "bar" tag doesn't have an int attribute named "X" (only the first "bar" tag does)
+		assertThrown!SDLangRangeException( root.getAttribute!string("*:bar", "attrNS:X") );
+		assert( root.getAttribute!string("*:bar", "attrNS:X", "Default") == "Default" );
+		
 	}
 	
-	private T getAttributeImpl(T)(string namespace, string tagName, string attrNamespace, string attrName)
+	private T getAttributeImpl(T)(string tagNamespace, string tagName, string attrNamespace, string attrName)
 	if(isValueType!T)
 	{
-		return getAttributeImpl!T(namespace, tagName, attrNamespace, attrName, T.init, false);
+		return getAttributeImpl!T(tagNamespace, tagName, attrNamespace, attrName, T.init, false);
 	}
 	
-	private T getAttributeImpl(T)(string namespace, string tagName, string attrNamespace,
+	private T getAttributeImpl(T)(string tagNamespace, string tagName, string attrNamespace,
 	string attrName, T defaultValue, bool useDefaultValue=true)
 	if(isValueType!T)
 	{
-		if(namespace !in _tags)
+		if(tagNamespace !in _tags)
 		{
 			if(useDefaultValue)
 				return defaultValue;
 			else
-				throw new SDLangRangeException("No attributes found in namespace '"~namespace~"'");
+				throw new SDLangRangeException("No attributes found in namespace '"~tagNamespace~"'");
 		}
 
-		if(tagName !in _tags[namespace] || _tags[namespace][tagName].length == 0)
+		if(tagName !in _tags[tagNamespace] || _tags[tagNamespace][tagName].length == 0)
 		{
 			if(useDefaultValue)
 				return defaultValue;
 			else
-			{
-				if(namespace == "*")
-					throw new SDLangRangeException("Can't find tag '"~tagName~"' in any namespace");
-				else
-					throw new SDLangRangeException("Can't find tag '"~tagName~"' in namespace '"~namespace~"'");
-			}
+				throw new SDLangRangeException("Can't find tag matching '"~FullName.combine(tagNamespace, tagName)~"'");
 		}
 
-		auto foundTag = _tags[namespace][tagName][$-1];
+		auto foundTag = _tags[tagNamespace][tagName][$-1];
 		
 		if(attrNamespace !in foundTag._attributes || attrName !in foundTag._attributes[attrNamespace])
 		{
@@ -1306,24 +1210,14 @@ class Tag
 				return defaultValue;
 			else
 			{
-				if(namespace == "*")
-				{
-					throw new SDLangRangeException(
-						"The last tag named '"~tagName~"' in all namespaces doesn't "~
-						"have any attributes '"~FullName.combine(attrNamespace, attrName)~"'"
-					);
-				}
-				else
-				{
-					throw new SDLangRangeException(
-						"The last tag named '"~tagName~"' in namespace '"~namespace~"' doesn't "~
-						"have any attributes '"~FullName.combine(attrNamespace, attrName)~"'"
-					);
-				}
+				throw new SDLangRangeException(
+					"The last tag matching '"~FullName.combine(tagNamespace, tagName)~"' doesn't "~
+					"have any attributes '"~FullName.combine(attrNamespace, attrName)~"'"
+				);
 			}
 		}
 
-		foreach(attr; foundTag._attributes[""][attrName])
+		foreach(attr; foundTag._attributes[attrNamespace][attrName])
 		{
 			if(attr.value.type == typeid(T))
 				return attr.value.get!T();
@@ -1333,20 +1227,10 @@ class Tag
 			return defaultValue;
 		else
 		{
-			if(namespace == "*")
-			{
-				throw new SDLangRangeException(
-					"The last tag named '"~tagName~"' in all namespaces doesn't "~
-					"have an attribute '"~FullName.combine(attrNamespace, attrName)~"' of type "~T.stringof
-				);
-			}
-			else
-			{
-				throw new SDLangRangeException(
-					"The last tag named '"~tagName~"' in namespace '"~namespace~"' doesn't "~
-					"have an attribute '"~FullName.combine(attrNamespace, attrName)~"' of type "~T.stringof
-				);
-			}
+			throw new SDLangRangeException(
+				"The last tag matching '"~FullName.combine(tagNamespace, tagName)~"' doesn't "~
+				"have an attribute '"~FullName.combine(attrNamespace, attrName)~"' of type "~T.stringof
+			);
 		}
 	}